+++ ea-factories/src/index.ts
@@ -1,0 +1,5 @@
+import * as balance from './factories/balance'
+import * as price from './factories/price'
+
+export { balance }
+export { price }